requirements.txt
python-telegram-bot[job-queue]==21.9
aria2p==0.12.0
python-dotenv==1.0.0
validators==0.22.0
setuptools>=65.5.1 